refactor(frontend): consolidate import/export into two settings tabs
The Settings sidebar listed four separate rows (Import, Export, Calendar
Import, Calendar Export) — a data-type × operation grid flattened into a
list that would grow to six rows once Contacts import/export is added.

Collapse to two sidebar rows (Import, Export). Each renders a container
with a Mail | Calendar segmented control (TabButtons) that swaps in the
relevant panel. Rename the mail-specific panels to MailImportSettings /
MailExportSettings and drop the now-redundant per-panel headings, since
the container supplies the heading and the data type is shown by the
segmented control.

Adding Contacts later is a two-line change per container (one button +
one component-map entry); no sidebar changes needed.
